requirements.txt
absl-py==2.3.1
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.6.15
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
cmake==4.2.0
colorama==0.4.6
contourpy==1.3.1
cryptography==46.0.4
cycler==0.12.1
dash==3.3.0
defusedxml==0.7.1
dilib==1.1.0
easyocr==1.7.2
face-recognition-models==0.3.0
fastapi==0.128.0
filelock==3.18.0
Flask==3.1.0
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.55.0
fpdf==1.7.2
fpdf2==2.8.3
fsspec==2025.5.1
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.189.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
google-pasta==0.2.0
googleapis-common-protos==1.72.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
h5py==3.14.0
httplib2==0.31.2
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
itsdangerous==2.2.0
jax==0.7.1
jaxlib==0.7.1
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keras==3.11.2
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
lxml==6.0.1
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.3
MouseInfo==0.1.3
mpmath==1.3.0
mysql-connector-python==9.3.0
namex==0.1.0
narwhals==1.45.0
nest-asyncio==1.6.0
networkx==3.5
ninja==1.11.1.4
nltk==3.9.1
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
orjson==3.11.7
packaging==24.2
pandas==2.3.0
patsy==1.0.1
pillow==11.0.0
plotly==6.3.0
proto-plus==1.27.1
protobuf==5.29.5
pyarrow==20.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pyclipper==1.3.0.post6
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pydicom==3.0.1
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==2.0.1
pyparsing==3.2.0
PyPDF2==3.0.1
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-pptx==1.0.2
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
referencing==0.36.2
regex==2024.11.6
reportlab==4.3.1
requests==2.32.4
retrying==1.4.2
rich==14.1.0
rpds-py==0.26.0
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentencepiece==0.2.1
shapely==2.1.1
six==1.16.0
smmap==5.0.2
sounddevice==0.5.3
starlette==0.50.0
statsmodels==0.14.5
streamlit==1.46.1
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.6.11
toml==0.10.2
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.40.0
watchdog==6.0.0
Werkzeug==3.1.3
wrapt==1.17.3
xlsxwriter==3.2.5
zipp==3.23.0
